% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{add_sera}
\alias{add_sera}
\title{Workaround to add SERA measure results to the}
\usage{
add_sera(x, form)
}
\arguments{
\item{form}{}
}
\value{

}
\description{
Workaround to add SERA measure results to the
}
