% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{ove.perc}
\alias{ove.perc}
\title{Tune parameter for oversampling percentage}
\usage{
ove.perc(range = c(1, 4), trans = NULL)
}
\arguments{
\item{range}{Vector of size 2 with the bounds for the range of the parameter. By default, ]1,4]}

\item{trans}{Transformation function.}
}
\value{

}
\description{
Tune parameter for oversampling percentage
}
\examples{
ove.perc()
ove.perc() \%>\% value_sample(5)
}
