% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{sera.data.frame}
\alias{sera.data.frame}
\title{Data.frame implementation of the SERA measure}
\usage{
\method{sera}{data.frame}(data, truth, estimate, ph, na_rm = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Data.frame implementation of the SERA measure
}
