% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluation.R
\name{sera_vec}
\alias{sera_vec}
\title{Vector implementation of the SERA measure}
\usage{
sera_vec(truth, estimate, ph, na_rm = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Vector implementation of the SERA measure
}
\examples{
data("solubility_test")
sera_vec(
truth = solubility_test$solubility,
estimate = solubility_test$prediction,
ph = IRon::phi.control(solubility_test$solubility)
)
}
