% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandOverClassif.R
\name{step_RandOverClassif}
\alias{step_RandOverClassif}
\title{Random Oversampling for Imbalanced Classification Tasks}
\usage{
step_RandOverClassif(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  ove.perc = 0.5,
  perc.list = NULL,
  C.perc = NULL,
  target = NULL,
  skip = TRUE,
  id = rand_id("RandOverClassif")
)
}
\arguments{
\item{recipe}{Recipe object}

\item{...}{Other parameters}

\item{role}{Not used by this step since no new variables are created.}

\item{ove.perc}{Oversampling percentage}

\item{perc.list}{Oversampling percentage list describing the values for each class}

\item{C.perc}{Final list of percentages}

\item{target}{Name of the target variable}

\item{skip}{dsa}

\item{id}{dsa}
}
\value{

}
\description{
dsa dsa
}
\examples{
library(tidymodels)
library(modeldata)

data(hpc_data)

hpc_data0 <- hpc_data \%>\%
  select(-protocol, -day)

down_rec <- recipe(class ~ ., data = hpc_data0) \%>\%
  step_RandOverClassif(class, ove.perc = 2) \%>\%
  prep()

tidy(down_rec, number=1)

training <- down_rec \%>\%
  bake(new_data = NULL)

table(hpc_data0$class)
table(training$class)

folds <- vfold_cv(hpc_data0, v = 5)

tune_rec <- recipe(class ~ ., data = hpc_data0) \%>\%
  step_RandOverClassif(class, ove.perc = tune())

lin_mod <-
  decision_tree() \%>\%
  set_mode("classification") \%>\%
  set_engine("rpart")

wf <- workflow() \%>\%
  add_recipe(tune_rec) \%>\%
  add_model(lin_mod)

lambda_grid <- grid_random(ove.perc(),size = 3) #strange bug

res <- tune_grid(wf,
                 resamples = folds,
                 grid = lambda_grid,
                 metrics = eval_metrics_classification)

estimates <- collect_metrics(res)
estimates

show_best(res, metric = "accuracy")
show_best(res, metric = "bal_accuracy")
show_best(res, metric = "roc_auc")
show_best(res, metric = "f_meas")

}
