% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandUnderClassif.R
\name{step_RandUnderClassif}
\alias{step_RandUnderClassif}
\title{Random Undersampling for Imbalanced Classification Tasks}
\usage{
step_RandUnderClassif(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  und.perc = 0.5,
  perc.list = NULL,
  C.perc = NULL,
  target = NULL,
  skip = TRUE,
  id = rand_id("RandUnderClassif")
)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
Random Undersampling for Imbalanced Classification Tasks
}
\examples{
library(tidymodels)
library(modeldata)
library(IDL)
data(hpc_data)

hpc_data0 <- hpc_data \%>\%
  select(-protocol, -day)

up_rec <- recipe(class ~ ., data = hpc_data0) \%>\%
  step_RandUnderClassif(class, und.perc = 0.5) \%>\%
  prep()

tidy(up_rec, number=1)

training <- up_rec \%>\%
  bake(new_data = NULL)

table(hpc_data0$class)
table(training$class)

folds <- vfold_cv(hpc_data0, v = 5)

tune_rec <- recipe(class ~ ., data = hpc_data0) \%>\%
  step_RandUnderClassif(class, und.perc = tune())

lin_mod <-
  decision_tree() \%>\%
  set_mode("classification") \%>\%
  set_engine("rpart")

wf <- workflow() \%>\%
  add_recipe(tune_rec) \%>\%
  add_model(lin_mod)

lambda_grid <- grid_random(und.perc(),size = 3)

res <- tune_grid(wf,
                 resamples = folds,
                 grid = lambda_grid,
                 metrics = eval_metrics_classification)

estimates <- collect_metrics(res)
estimates

show_best(res, metric = "accuracy")
show_best(res, metric = "bal_accuracy")
show_best(res, metric = "roc_auc")
show_best(res, metric = "f_meas")

}
