% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{und.perc}
\alias{und.perc}
\title{Tune parameter for undersampling percentage}
\usage{
und.perc(range = c(0, 1), trans = NULL)
}
\arguments{
\item{range}{Vector of size 2 with the bounds for the range of the parameter. By default, ]0,1[}

\item{trans}{Transformation function.}
}
\value{

}
\description{
Tune parameter for undersampling percentage
}
\examples{
und.perc()
und.perc() \%>\% value_sample(5)
}
